refactor : 기존 JpaRepository를 의존하는 레포의 이름 변경
해당 레포는 SpringDataJpa에 강하게 의존하고 있다.
그에 맞게 이름을 변경한다.
