+++ src/main/java/com/example/quaterback/login/repository/refresh/SpringDataRefreshRepository.java
@@ -1,4 +1,4 @@
-package com.example.quaterback.login.repository;
+package com.example.quaterback.login.repository.refresh;
 
 import com.example.quaterback.login.entity.RefreshEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
