+++ src/main/java/com/example/quaterback/login/repository/user/SpringDataJpaUserRepository.java
@@ -3,7 +3,7 @@ package com.example.quaterback.login.repository;
 import com.example.quaterback.login.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface SpringDataJpaUserRepository extends JpaRepository<UserEntity, Integer> {
 
     Boolean existsByUsername(String username);
